# estoque/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Servico de Estoque",
    "version": "2.0.0"
  },
  "paths": {
    "/produtos/{produto_id}/disponibilidade": {
      "get": {
        "summary": "Verificar disponibilidade de um produto",
        "operationId": "verificar_disponibilidade_produtos__produto_id__disponibilidade_get",
        "parameters": [
          {
            "name": "produto_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Produto Id"
            }
          },
          {
            "name": "quantidade",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Quantidade"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisponibilidadeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/simular-falha": {
      "post": {
        "summary": "Ativar/desativar falha simulada (demo Circuit Breaker)",
        "operationId": "simular_falha_admin_simular_falha_post",
        "parameters": [
          {
            "name": "ativo",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Ativo"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/": {
      "get": {
        "summary": "Status do servico de estoque",
        "operationId": "status__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "DisponibilidadeResponse": {
        "properties": {
          "produto_id": {
            "type": "string",
            "title": "Produto Id"
          },
          "quantidade_solicitada": {
            "type": "integer",
            "title": "Quantidade Solicitada"
          },
          "disponivel": {
            "type": "boolean",
            "title": "Disponivel"
          },
          "quantidade_estoque": {
            "type": "integer",
            "title": "Quantidade Estoque"
          }
        },
        "type": "object",
        "required": [
          "produto_id",
          "quantidade_solicitada",
          "disponivel",
          "quantidade_estoque"
        ],
        "title": "DisponibilidadeResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}